[dev mode] Move new AgentMetrics check to checks.d
